requirements.txt
django
#static file uchun
#hosting uchun 
gunicorn==21.2.0
psycopg2-binary==2.9.9
# django-environ==0.11.2
pillow==10.4.0
#back end uchun
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf-yasg==1.21.7